+++ packages/manager/src/components/TransferDisplay/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './TransferDisplay';
